requirements.txt
litellm 
langfuse 
langchain-groq 
langchain-huggingface
langchain_chroma
chromadb
rank_bm25
unstructured
pydantic 
instructor
stripe
colorama
python-dotenv